Add tests for onboard component
Add react-test-renderer as dependency to test react components.
Add new prop `onComplete` to `OnboardContainer` which is a callback to
be triggered when the onboard process is complete.
Move stateless components `AdminForm`, `SMTPForm` and `StorageForm` to
their own files.
Write tests for user interactions in `Onboard.test.js`.
